apply estimate logic about IISS
